requirements.txt
pyrogram==2.0.106
tgcrypto==1.2.5
motor==3.3.2
aiohttp==3.9.1
python-dotenv==1.0.0
pymongo==4.6.3
asyncpg>=0.29.0
aiosqlite
cachetools
pytz
catboxpy
aiofiles
psutil>=5.9.0
redis>=4.5.0
